No more caching or continous main loop
Aaaaaand there goes most of the work that went into this...
Caching was a cool idea, as was the implementation of the continuous
main loop. But this is going to be hosted on Heroku, where a sleeping
app still counts as active and will eat up my hours, so I need to use a
scheduler there instead. The cache wouldn't have persisted betwen dyno
executions, anyway.

Signed-off-by: Zach Palumbo <zach@palumbo.io>