Remove dependency on Musixmatch
Top tracks are now obtained from the Billboard RSS feed.

The top 40 listing on Musixmatch was not actually the Billboard top
listing; it was the top-searched lyrics on that site. Musixmatch was
also a commercial API, and while the free rate limits weren't really an
issue in current usage, it's nice to only use free stuff now.